+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 2020-10-16
+- added support for huh_filter fields
+- added util for retrieving autocompletejs configuration for a field 
+- added retrieval of value from autocompletejs response if keys are set in field config
+
 ## [0.1.3] - 2020-09-29
 - fixed focus and blur of results_container to hide results if input looses focus